Adding Alter test under unit/adapter.
