+++ lib/waterline/adapter/sync/strategies/alter.js
@@ -96,7 +96,12 @@ module.exports = function(cb) {
       //
       // From this point forward, we must be very careful.
       //
-      backupData = _.cloneDeep(existingData);
+      backupData = _.cloneDeep(existingData, function dealWithBuffers(val) {
+            if (val instanceof Buffer) {
+                return val.slice();
+            }
+      });
+
 
       // Check to see if there is anything obviously troublesome
       // that will cause the drop and redefinition of our schemaful
